moved home folder, edit typos
