+++ frontend/src/pages/Home/NavbarElements.js
